+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+# 0.3
+
+- Use regular expresions in some cases.
+- Add 'éxitos' and 'exitos'.
+
+# 0.2
+
+- Add 'emoprender'.
+
 # 0.1
 
 - Initial release.